Critical Fix: std::vector::size must be explicitly cast to uint32_t
Otherwise on x64 architecture the std::size_t is uint64_t and
BytestreamOut would emit 4 extra bytes that are not expected during
deserialization.

+++ Libraries/LevelD/Modules/ModuleMesh.cpp
@@ -68,7 +68,7 @@ void ModuleMesh_v2::deserialize(BytestreamIn &bin, LevelD &lvld) const {
 void ModuleMesh_v3::serialize(BytestreamOut &bout, const LevelD &lvld) const {
 	bout << lvld.mesh.tileWidth << lvld.mesh.tileHeight;
 	bout << lvld.mesh.layerWidth << lvld.mesh.layerHeight;
-	bout << lvld.mesh.layers.size();
+	bout << uint32_t(lvld.mesh.layers.size());
 
 	for (auto &layer : lvld.mesh.layers) {
 		std::vector<uint16_t> dataout(lvld.mesh.layerWidth * lvld.mesh.layerHeight);
